% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miReactC.R
\name{miReactC}
\alias{miReactC}
\title{Calculate miRNA activity using C++}
\usage{
miReactC(sco, n.cores = 1, verbose = T)
}
\arguments{
\item{sco}{A list containing indexes (var), motif p-values (pval.mat), motif counts (motifCounts), and alpha (runparameters$alpha)}

\item{verbose}{Show progress (default=T)}
}
\value{
A matrix with the estimated miRNA activities
}
\description{
Calculates miRNA activity using C++
}
