% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miReactC.R
\name{prepareData}
\alias{prepareData}
\title{Prepare data}
\usage{
prepareData(
  sco,
  seqs.path = "./seqs/hs.utr3.seqs.rds",
  seqlist.path = "./seqs/hs.utr3.seqlist.rds",
  patterns.path = "./motif.models/patterns.7mer.Rdata",
  pval.path = "./motif.models/hs.seqXmot.utr3_mrs_7mer.rds",
  counts.path = "./motif.models/hs.seqXmot.counts.utr3_mrs_7mer.rds",
  verbose = T
)
}
\arguments{
\item{sco}{List containing raw counts (exp) and run parameters (runparameters)}

\item{seqs.path}{Path to motif counts (default="./hs.seqXmot.counts.utr3_mrs_7mer.rds")}

\item{seqlist.path}{Path to motif counts (default="./hs.seqXmot.counts.utr3_mrs_7mer.rds")}

\item{patterns.path}{Path to motif counts (default="./hs.seqXmot.counts.utr3_mrs_7mer.rds")}

\item{pval.path}{Path to motif counts (default="./hs.seqXmot.counts.utr3_mrs_7mer.rds")}

\item{counts.path}{Path to motif counts (default="./hs.seqXmot.counts.utr3_mrs_7mer.rds")}

\item{verbose}{Show progress (default=T)}
}
\value{
A list with RNA count data, sequences, motif p-values, motif counts, and indexes
}
\description{
Prepare data for miReactC calculations
}
